+++ tests/util/linkComponents.js
@@ -8,7 +8,7 @@ describe('linkComponentsFunctions', () => {
     it('returns a default map of components', () => {
       const context = {};
       assert.deepStrictEqual(linkComponentsUtil.getLinkComponents(context), new Map([
-        ['a', 'href'],
+        ['a', ['href']],
       ]));
     });
 
@@ -19,6 +19,10 @@ describe('linkComponentsFunctions', () => {
           name: 'Link',
           linkAttribute: 'to',
         },
+        {
+          name: 'Link2',
+          linkAttribute: ['to1', 'to2'],
+        },
       ];
       const context = {
         settings: {
@@ -26,9 +30,44 @@ describe('linkComponentsFunctions', () => {
         },
       };
       assert.deepStrictEqual(linkComponentsUtil.getLinkComponents(context), new Map([
-        ['a', 'href'],
-        ['Hyperlink', 'href'],
-        ['Link', 'to'],
+        ['a', ['href']],
+        ['Hyperlink', ['href']],
+        ['Link', ['to']],
+        ['Link2', ['to1', 'to2']],
+      ]));
+    });
+  });
+
+  describe('getFormComponents', () => {
+    it('returns a default map of components', () => {
+      const context = {};
+      assert.deepStrictEqual(linkComponentsUtil.getFormComponents(context), new Map([
+        ['form', ['action']],
+      ]));
+    });
+
+    it('returns a map of components', () => {
+      const formComponents = [
+        'Form',
+        {
+          name: 'MyForm',
+          formAttribute: 'endpoint',
+        },
+        {
+          name: 'MyForm2',
+          formAttribute: ['endpoint1', 'endpoint2'],
+        },
+      ];
+      const context = {
+        settings: {
+          formComponents,
+        },
+      };
+      assert.deepStrictEqual(linkComponentsUtil.getFormComponents(context), new Map([
+        ['form', ['action']],
+        ['Form', ['action']],
+        ['MyForm', ['endpoint']],
+        ['MyForm2', ['endpoint1', 'endpoint2']],
       ]));
     });
   });